Fix misguided strlen() in Chop() and StripTrailingNewline().
They check strlen() doesn't exceed max_length, which callers typically
pass as the buffer size; either this test is a waste of time or, when
it might fail, it'll involve reading past the end of the buffer to
discover the error - which is bad.  So use strnlen() instead, this is
what it's for.  Don't bother special-casing empty string; and re-use
the value strnlen has returned to us instead of calling again at the
start of the search for the last non-space.  Only '\0' out the first
trailing space, rather than filling the tail with '\0' over each
space.  (For "space" read isspace() in Chop() and '\n' in
StripTrailingNewline().)  Structure the two functions the same way.