Redmine #5557: Fix files promises having wrong status report.
In functions where we invoke sub promises, we need to start with the
assumption that the promise will be skipped, instead of it being a
kept promise. The reason is that the skipped status has the lowest
priority, so if we end up skipping a promise, the status will still be
kept if that's what we started with.

(cherry picked from commit 9f36a7d053e7038a19b3146ca4f934bf11a6ca8b)