Redmine #5496: Fix enterprise_extension_test for ARM processors.
Because of function argument padding, the "char" argument is not
enough to trigger the stack inconsistency we are looking for in the
test. This also highlights that not all forms of stack corruption will
be detected when loading the plugin, it is just a safeguard.
(cherry picked from commit ecb5833019794cec2f03f349530699e2274b01b0)