Prevent user home dir creation on systems that try to enforce it.
Some systems (CentOS 6.4, for instance) have the default policy to
create the home directory whenever "useradd" is called. But we want
the home_bundle attribute to do this, so we need to prevent it.The
parameter to prevent this is "-M", but it is not available everywhere,
so we call useradd with the --help parameter first to see if it is in
the list of options.

Redmine #5781.

(cherry picked from commit 9ce44be5c5f39cb90ab43a1ca295c5aa67a0afa8)