Clean-up in verify_packages.c
Split some long lines.  Combined a decl with its init to move the
latter out of a loop's condition.  Use strlcpy in preference to
snprintf(,,"%s",); and give the true buffer size.  Reworked a for(;;)
loop that started with a block that break;s to a while loop with the
block moved after the loop; and asserted the the assumption a comment
was articulating.