Clean-up in libpromises/locks.c
A few ?: expressions can make code a lot less verbose.  Use judicious
assignments to skip the need to copy data into a buffer and reduce
duplication within #if-ery.  Simplify return from a bool function.
Deliver return values via locals so as to avoid duplicating clean-up.
Combine declarations and initializers, avoid assignments in
conditions.  Split some long lines.

Split a long line, use memcpy in place of sequential assignments.
Test the return from sscanf() instead of assuming a time of 13:00:43
is not a time, just because one of its components has the value you
used to initialize the variable into which you read it.  Don't
arbitrarily hard-code 15 as the length of a "day number plus time"
string when sprintf can actually tell us how long it was.  Comment on
further ways that RemoveDates() might be improved.

When keeping track of the size of copied data, to avoid overflowing a
buffer, exploit that knowledge of size to save making strcat traverse
all of the data repeatedly in order to append after it; and count the
dots in the data, as well as the texts between them, so that you do in
fact track the size, not a lower bound on it.