Leak-fix: free db_handles[*].filename in CloseAllDBExit().
The filename was created by xstrdup() via DBIdToPath() and
MapNameCopy(), so we have a duty to free() it eventually.

Cowardice: I'm only doing it, for now, if all DB users finished, since
I don't know what pain it might cause for one of them that's still
running ... but I left a TODO comment to look into that.