Rework KeepServerRolePromise().
I noticed a spurious continue (it was a no-op), so looked closely.
The inner loop tested a condition that didn't depend on any inner loop
variable, so I moved that outside the loop.  Different branches of the
switch on rval's type tested lval against various things; it makes
more sense to test the lval first and use that to decide what the
rval's type should be.  In particular, even if an "authorize" has a
value of the wrong type, we should report it even if it's currently
not active due to depending on a class that's not defined.  Moved a
function call outside a loop condition, made a variable more local.
Simplify loop by traversing Seq in reverse order.