Redmine #4241: Depr. INTERRUPTED promise status that was never used.
There was some code for setting classes based on this promise status,
but it was never implemented in the syntax, so it was impossible to
set a class based on it. Since we don't really need to introduce any
more promise outcomes, we alias the INTERRUPTED status to the "failed"
classes attribute.

The actual status is however left intact since it has been available
in reporting.