Redmine #4241: Make sure failed status carries through to the edit end.
Before, we would happily assume that if number of edits were zero,
this would mean that the promise was kept, however it could also be
an error. Fixed by checking the promise status in FinishEditContext().

Note the 017.cf test case, which was testing the inverse condition of
what it claimed to test.