Don't use "sudo -n" option, which is not supported by RHEL4.
It was added to avoid a password prompt when running tests. Jenkins
disables TTY access, so it should not be a problem there regardless,
but it may lead to cf-agent hanging on a password prompt when running
tests locally if sudoers is not set up correctly.

(cherry picked from commit 9a15c73ee7b2a494f559eeeb06fa38776f1bd564)