Redmine#5667 - address ps -axo shift on FreeBSD 10.0-RELEASE and later
As discussed and documented in Redmine#5667 , there's an inevitable creep to the right going on in FreeBSD 10 and later which has been demonstrated to break bootstrapping in a severe fashion. Going with the most elegant fix, which is to just say 'auxw' in line with Darwin, which is closest for these purposes. Output and column starts do not change, just guarantees getting complete lines regardless of terminal width. Quick test shows no impact.
Note that columns with -axo DO variably shift now on 10 and later, so systems with long runtimes trying to bootstrap are likely to encounter the described bug.
(cherry picked from commit 50df6902f08648fb82a5aa9758af17bad38d27a2)

Conflicts:
	libpromises/classes.c