Fixes for inappropriate formats in a test.
unsigned long isn't size_t.

It made no sense to use an intptr_t for the counter of a loop bounded
by an int.  The counter is cast to void*, but casting it via intptr_t
(to silence warnings about size-mangling casts) makes more sense than
making the counter actually be an intptr_t.  Likewise, when collecting
the int back from a void*, cast via intptr_t but use an int.