Redmine #5973: Lower verbosity level of harmless message.
Signals are normal in cf-hub's lifetime, since the parent process may
kill children that are taking too long to complete (for example
because of saturated network). Detecting this is more appropriately
done in Mission Portal than an error message on the console.

Other than cf-hub, all agents are supposed to handle signals
gracefully, so it should not be displayed as an error.