Handle socket()-failure like all others.
The child_cycle() test code neglected to check whether socket()
succeeded.  When it fails, handle it the same as all other failures.
Also, in other failures, don't bother saving write()'s return to a
variable when we're not going to do anything with it; and don't leave
space between a function's name and its parameter list.