Bug in makerule - and make it compatible with if/ifvarclass.
Function was declared incorrectly and would not match in
if/ifvarclass. Also it must return false in order to be compatible
with these cases, not merely fail as previously argued. This is the
only way to make class functions consistent

Downgrade the level of a Log() message.

(cherry picked from commit e3f11abf91a07a78439594ab32819c0d0514b3ee)