Use xsnprintf() when there is no need to check the return value.
NOTE: There is still a bunch of unchecked snprintf() usecases, but it's
not appropriate to convert them. For example, similar to the following
are found all over the place:

snprintf(buf, CF_BUFSIZE, "%s/%s", ...);

This might overflow if strings are large enough. No matter how uncommon
this is, the proper thing is to manually check snprintf() return value
and act accordingly (e.g. promise not kept).