redmine6147: the other fix.
It's possible for the handling of a connection to close that
connection; it's typically better to leave the HandleConnection()
level of code to deal with that, but this level may as well at least
cope gracefully if something *has* done so - provided it's suitably
marked the connection as having been closed.  So have DeleteConn()
actually pay attention to whether the socket is valid before closing
it.

(cherry picked from commit 01f36718c5a25993540b4fa7b7c3b8d2df296194)