Tidy up in NewConn and DeleteConn.
If you're going to calloc(), there's little point explicitly setting
the fields that get zero/false/NULL values.

Make clear that the getsockname() code in NewConn() only serves to
impose a failure, without using the data it finds.

Testing conn->ipaddr against NULL is senseless; it's guaranteed to
succeed, as ipaddr is an array member of the struct.  However, if the
string it contains is *empty*, we probably shouldn't be calling
DeleteItemMatching() on it !

Shuffled order of parts slightly for ease of reading.