Enable SIGINT (Ctrl-C) to kill testall instead of leaving it running.
It's quite convoluted, but basically it is a cascade of traps leading
up to killing the actual testall process. Note that it does not kill
any jobs that testall might have created. The hope is that these will
quit on their own in a short time. Basically each level you want to
kill requires a trap in the parent.