Use %jd and %ju instead of PRIdMAX and PRIuMAX.
The problem is that on Windows PRI[du]MAX expands to a
Windows-specific printf flag, but we are using printf wrappers, which
do not understand that flag. The 'j' flag on the other hand, is not
supported by Windows, but *is* supported by our wrapper, so it's safe
to use.

One notable exception from in the commit is a scanf which uses
PRIdMAX. scanf is not covered by our wrappers at the moment, and has
no replacement, so we need to leave that as it is. sprintf is not
covered either, but can (and should) be replaced with snprintf.