RedMine 6397 mitigation: use alternate column instead of empty.
In SplitProcLine(), we parse the line two ways; if the two don't
agree, we winge about it; but we use the first one anyway.  We see
some cases where Solaris ps gets a better answer from the second than
from the first; so use the second by default and fall back to the
first if it's empty.