Revert "Remove uninitialized buffers as source of entropy."
This reverts commit 570724e752cf131132bb88e65a031f700b0ac424.

The reason is that in a chroot testing environment, there isn't enough
entropy available to satisfy OpenSSL, so it will bail out. Reverting
is not the correct fix for the problem, but right now we need tests to
pass, so postponing this for later.

Conflicts:

	libpromises/crypto.c

(cherry picked from commit f0865c7682036ef78bb4b3ac27ea0d6a898b7958)