Avoid exposing libpromises' globals in libutils.
Add LoggingSetAgentType() and some state variables in logging.c by
which to let logging know the details it needs for Mark's recent
logging changes, without linker errors or cross-library linkage
complications.  Also avoid use of CF_BUFSIZE in logging.

Skip the VPREFIX when empty and move the space between it and the
timestamp (if present) to after it instead of before the timestamp so
that it won't abut the logging level when there's no timestamp and the
timestamp shall start at the beginning of the line if there's no
prefix.

Amend the timestamp test to take account of changes to logging -
mostly the fact that output is going to stdout now, not stderr, but
also make sure it commands the logging circumstances it assumes.