Change "Unacceptable model uncertainty" message and its log level.
It's not the most intelligible message - so reword it to give the
reader some chance of understanding it - and it's kinda spammy (albeit
more so than usual, lately, thanks to my recent bug in the parser).
Also add the name of the field we thought we were parsing, as that's
relevant.

(cherry picked from commit 30a310f4bf8476a76a30aae0559e9e1c2e359b9b)