Cope with ps's numeric (e.g. time) fields overflowing to the right.
This can happen, despite such fields normally being right-aligned,
when the field is too wide or some earlier field has overflowed, thus
pushing the affected field to the right.  This can push an internal
punctuator of the shunted field into the right-alignment position, so
adjust the test for "looks like a numeric field that may overspill
left" test as well as ajusting the "grow right" stanza to also happen
for numeric fields that have been shunted.

Prompted by overnight tests failing on a build machine which sees:
USER       PID  PPID  PGID %CPU %MEM  NI    SZ    VSZ   RSS NLWP STIME     ELAPSED     TIME COMMAND
jenkins  31941 31939 31939  1.1  0.2   0  9938  39752  4552    1 May20 92-21:17:55 1-02:07:14 sshd: jenkins@notty
(eliding a bunch of other lines), in which the TIME field's part
directly under the header, ":07:", starts and ends in colons, so was
not recognised as "numeric" (ends in digit); and needs to grow to the
right as well as left, despite being numeric (so trying to be
right-aligned).